Introduce "Feature Policy: vertical-scroll"
This CL introduces a feature which enables a page to ensure vertical
scrolling is always possible. Essentially, any frame which might need
to change the behavior of vertical scroll must have this feature on.

With this CL, frames that do not have "vertical-scroll" enabled will
receive "touchstart" and first "touchmove" events as non-cancelable
and are now allowed to have a CSS property for "touch-action" which
does not include "pan-y" (for any element inside). The future work for
this feature will limit the access to programmatic scrolling as well
so that techniques such as "element.scrollIntoView" could not be used to
extend vertical scrolling to outside the <iframe> bounds.

Bug: 611982
Change-Id: I991821496e30152b76d7cdc6bb295db650b60981