Set array length to 0 for disconnected worklet input
The spec says that if an input to an AudioWorkletNode is not
connected, the input should be represented as an input with a single
channel with a Float32Array of length 0.

Make it so by passing in nullptr for that input to the
AudioWorkletGlobalScope::Process so it can fill the arrays
appropriately.

Bug: 817145
Change-Id: I82407ed0a9fe84c5012333af8af27f4dd08d29b8